+++ app/src/main/java/com/battlelancer/seriesguide/shows/overview/SeasonsFragment.kt
@@ -29,7 +29,6 @@ import com.battlelancer.seriesguide.shows.episodes.EpisodeTools
 import com.battlelancer.seriesguide.shows.episodes.EpisodesActivity
 import com.battlelancer.seriesguide.ui.BaseMessageActivity
 import com.battlelancer.seriesguide.ui.dialogs.SingleChoiceDialogFragment
-import com.battlelancer.seriesguide.util.TextTools
 import com.battlelancer.seriesguide.util.ThemeUtils
 import org.greenrobot.eventbus.EventBus
 import org.greenrobot.eventbus.Subscribe
@@ -151,6 +150,7 @@ class SeasonsFragment() : Fragment() {
                     showSortDialog()
                     true
                 }
+
                 else -> {
                     false
                 }
@@ -161,10 +161,12 @@ class SeasonsFragment() : Fragment() {
     private val listOnItemClickListener = object : SeasonsAdapter.ItemClickListener {
         override fun onItemClick(v: View, seasonRowId: Long) {
             val intent = EpisodesActivity.intentSeason(seasonRowId, requireActivity())
-            ActivityCompat.startActivity(requireActivity(), intent,
+            ActivityCompat.startActivity(
+                requireActivity(), intent,
                 ActivityOptionsCompat
                     .makeScaleUpAnimation(v, 0, 0, v.width, v.height)
-                    .toBundle())
+                    .toBundle()
+            )
         }
 
         override fun onPopupMenuClick(v: View, seasonRowId: Long) {
@@ -176,22 +178,27 @@ class SeasonsFragment() : Fragment() {
                             onFlagSeasonWatched(seasonRowId, true)
                             true
                         }
+
                         R.id.menu_action_seasons_watched_none -> {
                             onFlagSeasonWatched(seasonRowId, false)
                             true
                         }
+
                         R.id.menu_action_seasons_collection_add -> {
                             onFlagSeasonCollected(seasonRowId, true)
                             true
                         }
+
                         R.id.menu_action_seasons_collection_remove -> {
                             onFlagSeasonCollected(seasonRowId, false)
                             true
                         }
+
                         R.id.menu_action_seasons_skip -> {
                             onFlagSeasonSkipped(seasonRowId)
                             true
                         }
+
                         else -> false
                     }
                 }
@@ -276,7 +283,12 @@ class SeasonsFragment() : Fragment() {
         val binding = binding ?: return
         val unwatched = result.unwatchedEpisodes
         binding.textViewSeasonsRemaining.text =
-            TextTools.getRemainingEpisodes(requireContext().resources, unwatched)
+            resources.getQuantityString(
+                R.plurals.remaining_of_total_episodes_plural,
+                unwatched,
+                unwatched,
+                result.totalEpisodes
+            )
         setWatchedToggleState(unwatched == 0)
         setCollectedToggleState(result.collectedAllEpisodes)
     }
@@ -309,10 +321,12 @@ class SeasonsFragment() : Fragment() {
                         onFlagShowWatched(true)
                         true
                     }
+
                     CONTEXT_WATCHED_SHOW_NONE_ID -> {
                         onFlagShowWatched(false)
                         true
                     }
+
                     else -> false
                 }
             }
@@ -347,10 +361,12 @@ class SeasonsFragment() : Fragment() {
                         onFlagShowCollected(true)
                         true
                     }
+
                     CONTEXT_COLLECTED_SHOW_NONE_ID -> {
                         onFlagShowCollected(false)
                         true
                     }
+
                     else -> false
                 }
             }
@@ -359,11 +375,13 @@ class SeasonsFragment() : Fragment() {
 
     private fun showSortDialog() {
         val sortOrder = SeasonsSettings.getSeasonSortOrder(requireContext())
-        SingleChoiceDialogFragment.show(parentFragmentManager,
-                R.array.sesorting,
-                R.array.sesortingData, sortOrder.index,
+        SingleChoiceDialogFragment.show(
+            parentFragmentManager,
+            R.array.sesorting,
+            R.array.sesortingData, sortOrder.index,
             SeasonsSettings.KEY_SEASON_SORT_ORDER, R.string.pref_seasonsorting,
-                "seasonSortOrderDialog")
+            "seasonSortOrderDialog"
+        )
     }
 
     private val onSortOrderChangedListener = OnSharedPreferenceChangeListener { _, key ->